feat(cmd): add audit
